Refactor P2
Program now compiles, fixed issue with duplicate m_CommandBuffers member variable in a struct and a singular vairable.
Also fixed bad logic for exiting the app, initially thought it was a weird interaction/bug between
glfw and arch/hyprland that made window not close but it was bad logic. However this brought up some
issues with vulkan cleanup as it has never been called before now. is yet to be fixed but it runs well
before we try and close it. wq